requirements.txt
django
djangorestframework
markdown
django-filter
djangorestframework-gis
django-cors-headers

